reverting last change

+++ app/models/harvest_event/collection_manager.rb
@@ -1,5 +1,4 @@
 class HarvestEvent
-  # TODO: rename
   class CollectionManager
     def self.sync(event)
       manager = new(event)
@@ -10,6 +9,8 @@ class HarvestEvent
       @event = event
     end
 
+    # TODO: something in this process is slow. Figure out what it is and speed
+    # it up!
     def sync
       EOL.log_call
       update_attributes
@@ -20,7 +21,7 @@ class HarvestEvent
       end
       # NOTE: because it's updated, it will ALSO reindex the collection in Solr:
       collection.fix_item_count
-      if @event.published?
+      if @event.published? && resource.preview_collection
         resource.preview_collection.users = []
         resource.preview_collection.destroy
       end
@@ -28,6 +29,7 @@ class HarvestEvent
 
     private
 
+    # TODO: break this method up.
     def add_items_collection
       EOL.log_call
       data = Set.new()
@@ -36,7 +38,7 @@ class HarvestEvent
         data << "'#{title}', 'DataObject', #{object.id}, #{collection.id}"
       end
       harvested_entries_not_already_in_collection.find_each do |entry|
-        data << "'#{entry.name.string}', 'TaxonConcept', "\
+        data << "'#{entry.name.string.gsub("'", "''")}', 'TaxonConcept', "\
           "#{entry.taxon_concept_id}, #{collection.id}"
       end
       EOL::Db.bulk_insert(CollectionItem,
@@ -118,15 +120,15 @@ class HarvestEvent
 
     def remove_collection_items_not_in_harvest
       EOL.log_call
-      CollectionItem.data_objects.
+      ids = CollectionItem.data_objects.
         select("collection_items.id").
         joins("LEFT JOIN data_objects_harvest_events dohe ON "\
           "(collection_items.collected_item_id = dohe.data_object_id "\
           "AND dohe.harvest_event_id = #{@event.id})").
         where(collection_id: collection.id).
         where("data_object_id IS NULL").
-        delete_all
-      CollectionItem.taxa.
+        pluck(:id)
+      ids += CollectionItem.taxa.
         select("collection_items.id").
         joins("LEFT JOIN (harvest_events_hierarchy_entries hehe JOIN "\
           "hierarchy_entries he ON (hehe.hierarchy_entry_id=he.id AND "\
@@ -134,7 +136,8 @@ class HarvestEvent
           "(collection_items.collected_item_id = he.taxon_concept_id)").
         where(collection_id: collection.id).
         where("hehe.hierarchy_entry_id IS NULL").
-        delete_all
+        pluck(:id)
+      CollectionItem.delete(ids)
     end
 
     def resource
@@ -160,7 +163,7 @@ class HarvestEvent
       max_id = collection.collection_items.maximum(:id) || 0
       yield
       collection.collection_items.where(["id > ?", max_id]).find_in_batches do |batch|
-        SolrCore::CollectionItem.reindex_items(
+        SolrCore::CollectionItems.reindex_items(
           CollectionItem.preload_collected_items(batch, richness_score: true))
       end
     end
